requirements.txt
pandas>=2.0.0
sqlparse>=0.4.4
streamlit>=1.35.0
groq>=0.11.0
pytest>=8.0.0
python-dotenv>=1.0.1
plotly>=5.0.0
openpyxl>=3.0.0
